DemaisAjustes

+++ MajuliA2/Models/Produto/ProdutoRequest.cs
@@ -9,7 +9,7 @@ namespace MajuliA2.Models.Produto
         public string Cor { get; set; }
 
         [Range(0, double.MaxValue)]
-        public decimal Valor { get; set; }
+        public int Valor { get; set; }
         [Required]
         public int CategoriaId { get; set; }
     }
